Get running off of external oscillator
Because PICs are PICs, there are some weird instructions that are
necessary in order to switch system clock sources, it's possible that
the code written to do this might break if optimizations are turned
on. So let's not do that.

Runs at 25MIPS now, but we can adjust that up to 120MIPS or down to
6ish