First commit where I successfully wrote to a file on the SD card
This commit writes "hello world!" into the file "/test.txt". The code
isn't very clean, but the functions are nicely broken out so we should
be fine. Also created a funciton error that lets me set one breakpoint
and be able to view the the error message. This error function will
someday handle the errors in some way.

The FAT32 library I'm using is the one from ultra-embedded
(http://ultra-embedded.com/fat_filelib/). It's released under GPL, so
I'm going to put the logger code under that as well.