Run all source files through astyle (source formatter)
Alex yells at me when my source code is poorly formatted. And I yell
at other people when their source is bad, so I guess it would be
hypocritical not to do this.

Most of the changes are k&r function declaration styles (the opening
curly brace goes on the /next/ line) and if( expansions to if (. Sorry
for the unbalanced parentheses :) :)