Add missing functionality in can_syslog.c
It now properly cycles through buffers when one buffer gets full, and it
properly logs full buffers onto the SD card. Also added an error enum
value for "all the buffers are full and we've been asked to store
another CAN message. Dropping it".

This commit does leave behind a TODO to get timestamping working
properly. Need to somehow hack in a get_uptime_us() function. That's for
a later date, for now it just writes 0xcafebabe as the timestamp of all
CAN functions. So that's fun.