Create sd card writing benchmark
This test writes 2000 CAN messages onto the SD card, and uses the LEDs
to display how much time that takes. Red LED goes on, we start
initializing the SD card, then blue LED goes on, then we write the 2000
messages, then both LEDs go off.

I'm going to use this to tell exactly how much improved each step in
this speedup process is. At present, with very naive SD card settings,
this benchmark takes an unknown amount of time to initialize the SD card
(sorry, wasn't measuring), and takes 8:28 to log those 2000 messages,
which is a bitrate of ((11 + 8*8)*2000 / (8*60 + 28)) = 295bps. Which is
about three orders of magnitude too slow. So let's speed that up a bit.