Increase SPI bitrate
When from 32MHz prescaled down 512:1 to 2:1. Didn't go all the way to
1:1 because the SD card doesn't seem to go that fast. You'd think that
this would cause the write to increase 256x, but it doesn't seem to.
Instead, it now takes 32 seconds to initialize the SD card, and 7:20 to
flush the data to the SD card. So it seems like we need some more stuff.