Increase CAN buffer size for logging to SD card
Go from 512 bits to 4096 bits. The thinking behind this is that there
will be fewer blocking calls to the SD card write functions, and that it
can write more blocks consecutively, which should reduce total runtime.
And in fact, it does: We're now at 32s to initialize the SD card, and
2:14 to write the 2000 files (1119bps). Now only 2 and a half orders of
magnitude to go.