Add SPI read/write buffer functions; remove delays
So there were 1ms delays in the SPI read/write functions taht don't seem
to be necessary. I've removed those, and that (understandably)
introduces a major speedup. I had to increase the number of messages to
log from 2000 to 20000, and it now takes 12 seconds to log all of them.
This is now a bitrate of 125kbps. This is probably sufficient for our
purposes; that's effectively 1600 CAN messages per second (call it 1000
just for simplicity and to account for other code that this isn't
calling), which is more than we ever /should/ be sending.

Yes, this isn't a full Mbps, which is the theoretical max bitrate of our
CAN bus. So yes, if you absolutely flood the bus, this will overrun the
logger. However, it's probably not worth running after that extra 10x
speedup, this is fine.