Change logger timestamp to real value, rev canlib
Increase logger version number to 0.2.0 because it now seems to actually
work at decent speeds. However, there is a minor issue where the output
files seem to be corrupted (a bunch of \200's show up in the txt file,
and I have no idea why). I'm assuming that it's an issue with the FAT
driver, TODO investigate that.