Add incrementing number to every CAN message logged
We have a small problem with the logger where sometimes it just misses
CAN messages. This is a thing that needs to be solved, but for now we
just need a way to verify whether or not a message is missed on the SD
card. The easiest way to determine this is to just put a unique
incrementing number before every logged CAN message. That way if we see
that message 500 is logged, then there's some corrupted stuff, then we
see message 505, we know that we missed 4 messages.