Remove benchmarking code and change init order
The benchmarking code shouldn't be in master, it now works properly.

Also, changed init order so that CAN is brought up before the SD card.
The SD card takes forever to init, so it's possible that we could miss
some CAN messages during that time. By initializing CAN first, those
messages get buffered and will be written to SD card once it's up and
we're in the main loop.