Fix bug with garbled output; ready for WDR2
The bug was caused by passing the FAT library more than 1024 bytes to
store to the SD card at once, it just quietly wrote the next 3000 odd
bytes as 0's (oops). So I reduced the maximum size we pass to it to 512
bytes, and the output looks good now.

I added in a main event loop: it now blinks the red LED, and it spits
out its status every 250ms onto the bus. Everything should be good to go
for WDR, but it isn't tested. I'll get to that tomorrow